refactor: Delete mockTRPCContext and mockData, and relocate mockStripeSubscription to stripe mock.
